Visual product manager/requirements.txt
Flask==2.3.4
Flask-Cors==3.0.10
SQLAlchemy==2.0.24
Pillow==10.1.0
imagehash==4.3.1
numpy==1.26.0
werkzeug==3.0.0
